pyproject.toml
# =================================
# Build System
# =================================
[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"


# =================================
# Project Metadata & Configuration
# =================================
[project]
name = "pdf-to-english-py"
version = "0.1.0"
description = "Prototype Mistral OCR pipeline for pdf-to-english"
readme = "README.md"
authors = [{ name = "Michelle" }]
requires-python = ">=3.14"
dependencies = [
    "gradio>=5.0.0",
    "markdown-it-py>=2.2.0",
    "mistralai>=1.10.1",
    "python-dotenv>=1.2.1",
    "weasyprint>=68.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]


# =================================
# Pyright Type Checking
# =================================
[tool.pyright]
# Type Checker is needed as Ruff is only a linter: https://docs.astral.sh/ruff/faq/
# For Cursor (Cursor Pyright extension), for VSCode (Microsoft Pylance)
# Cursor Pyright forks BasedPyright, some lint rules overlap with Ruff, causes dupe errors

typeCheckingMode = "standard" # Cursor Pyright default
pythonVersion = "3.14"

# Disable Basedpyright rules that duplicate Ruff
reportUnusedImport = false      # Ruff F401 handles it
reportUnusedVariable = false    # Ruff F841 handles it
reportDuplicateImport = false   # Ruff F811 handles it


# =================================
# Ruff Linter Configuration
# =================================
[tool.ruff]
# Strictest ruff settings
line-length = 88
target-version = "py314"

[tool.ruff.lint]
# Enable all available rules
select = ["ALL"]
# Only ignore rules that are incompatible or overly pedantic
ignore = [
    "D100",    # Missing docstring in public module
    "D104",    # Missing docstring in public package
    "D203",    # 1 blank line required before class docstring (conflicts with D211)
    "D213",    # Multi-line docstring summary should start at the second line (conflicts with D212)
    "COM812",  # Trailing comma missing (conflicts with formatter)
    "ISC001",  # Implicitly concatenated string literals on one line (conflicts with formatter)
    "T201",    # Print found (allow print statements)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"scripts/*" = [
    "INP001",  # Allow scripts to be standalone (not require __init__.py)
]
"tests/*" = [
    "D103",    # Missing docstring in public function
    "PLR2004", # Magic value used in comparison
    "S101",    # Use of assert detected
]

[tool.ruff.lint.pylint]
max-args = 5
max-branches = 12
max-returns = 6
max-statements = 50

[tool.ruff.lint.mccabe]
max-complexity = 10


# =================================
# Pytest Configuration
# =================================
[tool.pytest.ini_options]
markers = [
    "integration: tests that require external services (Mistral API)",
]
